requirements.txt
smartapi-python
pyotp
streamlit>=1.32.0
pandas>=2.0.0
numpy>=1.24.0
plotly>=5.18.0
requests
websocket-client
